Safari-611.1.10/LayoutTests/js/dom/domjit-accessor-document-element-changed-expected.txt

Test DOMJIT documentElement accessor works after the documentElement is changed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testNull(target, result) is false
PASS testNull(target, result) is true
PASS successfullyParsed is true

TEST COMPLETE
